lkjas

+++ Age of war/src/ar/edu/itba/game/CanAttack.java
@@ -4,6 +4,6 @@ public interface CanAttack {
 	public void attack(Attackable objective);
 	public int getAttackRange();
 	public Side getSide();
-	public double getX();
-	public double getWidth();
+	public float getX();
+	public int getWidth();
 }
