+++ Age of war/src/ar/edu/itba/game/Projectile.java
@@ -4,7 +4,6 @@ import java.awt.geom.Point2D;
 import java.util.Observable;
 
 import Observers.ProjectileObserver;
-import UserInterface.Textures;
 
 public class Projectile extends Observable{
 	private Element element;
@@ -31,8 +30,8 @@ public class Projectile extends Observable{
 		return new Point2D.Double(this.getMiddleX(), this.element.getY() + this.element.getHeight());
 	}
 	
-	public Projectile(double X, double Y, double velX, double velY, boolean gravity, int damage){
-		this.element = new Element((float)(X - 15) , (float)Y, (float)velX, (float)velY, 30, 50, true);
+	public Projectile(float X, float Y, float velX, float velY, boolean gravity, int damage){
+		this.element = new Element((X - 15) , Y, velX, velY, 30, 50, true);
 		this.damage = damage;
 		WorldManager.getInstance().getElements().add(this.element);
 		this.observer = new ProjectileObserver(this);
